rearrange and split pg code
